fix: address multiple code quality and safety issues across common and GUI layers
- parse-uri: guard all json_string_value() results against NULL in
  get_otpauth_uri; fix g_utf8_strdown() memory leak; fix get_otpauth_data
  to not pre-allocate a gcrypt secure buffer before passing to
  g_file_get_contents (caused a secure-memory leak and replaced the pointer
  with non-secure memory); remove spurious double g_set_error(); replace
  silent guint8-truncating casts for period/digits with validated range checks
- authpro: fix copy-paste bug where "issuer" key was read instead of "label"
  when building the Username export field; guard algo and type comparisons
  against NULL json_string_value() returns
- common: remove duplicate json_object_set() for "secret" in build_json_obj;
  replace C11-optional VLA stack buffers (salt, iv, tag) in
  get_data_from_encrypted_backup with g_autofree heap allocations
- db-common: unify hash type to guint32 in add_otps_to_db and check_duplicate
  (was mixing guint and guint32 across load_db and add_otps_to_db, causing
  potential UB on ILP64 platforms)
- aegis: add missing g_set_error() to five error paths in
  get_otps_from_encrypted_backup that returned NULL without setting the error;
  free salt and key_nonce on kdf_derive failure in export_aegis; guard type
  and algo fields against NULL before g_ascii_strcasecmp in parse_aegis_json_data
- password-cb: replace g_utf8_strlen+strncpy with strlen+memcpy when copying
  passwords into gcrypt secure buffers (character count vs byte count mismatch
  caused under-allocation for multibyte UTF-8 passwords)

+++ src/common/db-common.c
@@ -163,9 +163,9 @@ add_otps_to_db (GSList       *otps,
     for (guint i = 0; i < list_len; i++) {
         otp_t *otp = g_slist_nth_data (otps, i);
         obj = build_json_obj (otp->type, otp->account_name, otp->issuer, otp->secret, otp->digits, otp->algo, otp->period, otp->counter);
-        guint hash = json_object_get_hash (obj);
-        if (g_slist_find_custom (db_data->objects_hash, GUINT_TO_POINTER(hash), check_duplicate) == NULL) {
-            db_data->objects_hash = g_slist_append (db_data->objects_hash, g_memdup2 (&hash, sizeof (guint)));
+        guint32 hash = json_object_get_hash (obj);
+        if (g_slist_find_custom (db_data->objects_hash, GUINT_TO_POINTER((guint)hash), check_duplicate) == NULL) {
+            db_data->objects_hash = g_slist_append (db_data->objects_hash, g_memdup2 (&hash, sizeof (guint32)));
             db_data->data_to_add = g_slist_append (db_data->data_to_add, obj);
         } else {
             g_print ("[INFO] Duplicate element not added\n");
@@ -178,8 +178,8 @@ gint
 check_duplicate (gconstpointer data,
                  gconstpointer user_data)
 {
-    guint list_elem = *(guint *)data;
-    if (list_elem == GPOINTER_TO_UINT(user_data)) {
+    guint32 list_elem = *(guint32 *)data;
+    if (list_elem == (guint32)GPOINTER_TO_UINT(user_data)) {
         return 0;
     }
     return -1;
